backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
black==25.9.0
boto3==1.40.41
botocore==1.40.41
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cryptography==46.0.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.110.1
flake8==7.3.0
h11==0.16.0
httpx==0.25.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
jmespath==1.0.1
jq==1.10.0
markdown-it-py==4.0.0
mccabe==0.7.0
mdurl==0.1.2
motor==3.3.1
mypy==1.18.2
mypy_extensions==1.1.0
numpy==2.3.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pathspec==0.12.1
platformdirs==4.4.0
pluggy==1.6.0
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pyflakes==3.4.0
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.5.0
pytest==8.4.2
pytest-asyncio==0.21.1
pytest-cov==4.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
pytokens==0.1.10
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
s3transfer==0.14.0
s5cmd==0.2.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.25.0
watchfiles==1.1.0

# Machine Learning for Predictive Intelligence
scikit-learn==1.3.2
scipy==1.11.4
numpy==2.3.3

# Monitoring and Metrics
prometheus-client==0.19.0

# Distributed Tracing (OpenTelemetry)
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-aiohttp-client==0.42b0

# Secrets Management (optional - install as needed)
boto3==1.40.41  # AWS Secrets Manager
hvac==2.1.0  # HashiCorp Vault
azure-keyvault-secrets==4.8.0  # Azure Key Vault
azure-identity==1.15.0  # Azure Identity

# Production monitoring and logging
sentry-sdk[fastapi]==1.40.0
structlog==23.2.0

# Redis for caching (optional)
redis==5.0.1

# Rate limiting
slowapi==0.1.9

# API documentation
mkdocs==1.5.3
mkdocs-material==9.4.8

# MFA dependencies
pyotp==2.9.0
qrcode[pil]==7.4.2

# Development tools
pre-commit==3.6.0
bandit==1.7.5
